Fix build breaks from rebase.
